+++ src/components/MyPagePasswordForm/MyPagePasswordForm.tsx
@@ -3,6 +3,7 @@ import * as yup from "yup";
 import { yupResolver } from "@hookform/resolvers/yup";
 import { useForm } from "react-hook-form";
 import axios from "@/lib/axios";
+import UserService from "@/api/UserService";
 
 function PasswordForm() {
   // react-hook-form, yup 라이브러리를 통해 유효성 검사
@@ -31,42 +32,7 @@ function PasswordForm() {
   });
 
   const onSubmit = async (data) => {
-    // localStorage에서 accessToken의 비밀번호를 가져옴
-    const token = localStorage.getItem("accessToken");
-    const { password, newPassword } = data;
-
-    // // 입력한 현재 비밀번호와 localStorage에 저장된 비밀번호를 비교
-    // if (nowPassword !== storedPassword) {
-    //   alert("현재 비밀번호가 틀립니다.");
-    //   return;
-    // }
-    if (token) {
-      try {
-        // Axios를 사용하여 PUT 요청을 보내어 새로운 비밀번호로 업데이트
-        const response = await axios.put(
-          "/auth/password",
-          {
-            password,
-            newPassword,
-          },
-          {
-            headers: {
-              Authorization: `Bearer ${token}`,
-            },
-          }
-        );
-
-        if (response.status === 204) {
-          // 성공적으로 업데이트된 경우
-          alert("비밀번호가 업데이트되었습니다.");
-        } else {
-          // 업데이트에 실패한 경우
-          console.error("비밀번호 업데이트에 실패했습니다.");
-        }
-      } catch (error) {
-        console.error("비밀번호 업데이트 중 오류가 발생했습니다.", error);
-      }
-    }
+    UserService.updatePassword(data);
 
     // 비밀번호 변경 로직 실행
     // 여기서는 간단히 alert로 표시하도록 하겠습니다.
